# query/physicalplan: For special substring regexp filter run substr matching directly