Once a regexp filter is run it checks for .% at the beginning and end.
If it is true, then using bytes.Contains or strings.Contains is way more performant than running regexp.

Here's the result of running BenchmarkBinaryScalarRegexp:
```
name                      old time/op  new time/op  delta
BinaryScalarRegexp/=~-24   122ms ± 3%     1ms ±26%  -99.34%  (p=0.000 n=10+9)
BinaryScalarRegexp/!~-24   123ms ± 2%     1ms ±14%  -99.31%  (p=0.000 n=10+9)
```